// src/app/page/master/jenispegawai/jenispegawai.interface.ts
import {EventEmitter} from '@angular/core';

export interface JenisPegawai {
    kodeJenisPegawai?: string;
	namaJenisPegawai?: string;
	reportDisplay?: any;
	kodeExternal?: string;
	namaExternal?: string;
	kodeDetailKelompokPegawai?: any;
	kode?: any;
	version?: any;
	Version?: any;
	statusEnabled?: any;
	kodeKelompokPegawai?: any;
	label?:any;
	value?:any;
	namaKelompokPegawai?:any;
	kdKelompokPegawai?:any;
	id_kode?:any;
	
}